requirements.txt
pydantic==1.10.13
openai==0.28
httpx==0.27.2
azure-functions
tenacity==8.2.2